requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
trino==0.327.0
psycopg2-binary==2.9.10
PyYAML==6.0.2
python-dotenv
pydantic-settings